backend: refactor airtableBackup.js
