+++ backend/src/airtableBackup.js
@@ -1,21 +1,9 @@
-const { airtableBase, fetchAllRecords } = require("./airtable")
+// This file includes logic of backup of Airtable data (currently, only organizations) in Postgres. So far, it is used
+// only in addFirstTimeScrapingJobs.js, but, in principle, backup could also be done as part of periodic re-scraping:
+// see https://github.com/climatescape/climatescape.org/issues/110.
+
 const { setupTables } = require("./db/setupTables")
 const { executeCount, executeBulkInsertOrUpdate } = require("./db/pg")
-
-/**
- * @returns {Promise<Array<Object>>} array of orgs from Airtable
- */
-async function fetchAllOrgsFromAirtable() {
-  console.log("Fetching organizations from Airtable...")
-  const allOrgRecords = await fetchAllRecords(
-    airtableBase("Organizations").select()
-  )
-  // Cleans up irrelevant fields: references to the parent table, callback
-  // functions, etc.
-  return allOrgRecords.map(orgRecord => {
-    return { id: orgRecord.id, fields: orgRecord.fields }
-  })
-}
 
 /**
  * @param {Array<Object>} orgs - array or orgs from Airtable
@@ -57,12 +45,4 @@ async function backupOrgsInDb(orgs) {
   console.log(`Saved ${numOrgsAfter - numOrgsBefore} new organizations`)
 }
 
-/**
- * @returns {Promise<void>}
- */
-async function fetchAndBackupInDbAllAirtableOrgs() {
-  const orgs = await fetchAllOrgsFromAirtable()
-  await backupOrgsInDb(orgs)
-}
-
-module.exports = { backupOrgsInDb, fetchAndBackupInDbAllAirtableOrgs }
+module.exports = { backupOrgsInDb }
